Add puppet-blacksmith to Gemfile and Rakefile
Add puppet-blacksmith, a tool to make releasing to the forge easier.
The last time I tried this [1], I found that puppet-blacksmith needs
ruby 1.9 or greater.  I'm using the bundler platforms feature to get
bundler to only install the gem on ruby 1.9 or greater; the Rakefile
will only add the blacksmith rake tasks if it's present.

[1] https://github.com/gds-operations/puppet-graphite/commit/4e8b86200b7ff26